Secure university endpoints

+++ src/main/java/com/example/cms/university/UniversityService.java
@@ -2,16 +2,20 @@ package com.example.cms.university;
 
 import com.example.cms.page.Page;
 import com.example.cms.page.PageRepository;
+import com.example.cms.security.SecurityService;
 import com.example.cms.university.exceptions.UniversityException;
 import com.example.cms.university.exceptions.UniversityExceptionType;
 import com.example.cms.university.projections.UniversityDtoDetailed;
-import com.example.cms.university.projections.UniversityDtoForm;
+import com.example.cms.university.projections.UniversityDtoFormCreate;
+import com.example.cms.university.projections.UniversityDtoFormUpdate;
 import com.example.cms.university.projections.UniversityDtoSimple;
 import com.example.cms.user.User;
 import com.example.cms.user.UserRepository;
 import com.example.cms.validation.exceptions.BadRequestException;
+import com.example.cms.validation.exceptions.ForbiddenException;
 import com.example.cms.validation.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,19 +28,63 @@ public class UniversityService {
     private final UniversityRepository universityRepository;
     private final UserRepository userRepository;
     private final PageRepository pageRepository;
+    private final SecurityService securityService;
 
     @Autowired
-    public UniversityService(UniversityRepository universityRepository, UserRepository userRepository, PageRepository pageRepository) {
+    public UniversityService(UniversityRepository universityRepository, UserRepository userRepository,
+                             PageRepository pageRepository, SecurityService securityService) {
         this.universityRepository = universityRepository;
         this.userRepository = userRepository;
         this.pageRepository = pageRepository;
+        this.securityService = securityService;
+    }
+
+    public UniversityDtoDetailed getUniversity(Long id) {
+        // TODO: return only if visible
+        return universityRepository.findById(id).map(UniversityDtoDetailed::new).orElseThrow(NotFoundException::new);
     }
 
     public List<UniversityDtoSimple> getUniversities() {
+        // TODO: return only visible universities
         return universityRepository.findAll().stream().map(UniversityDtoSimple::new).collect(Collectors.toList());
     }
 
+    @Secured("ROLE_MODERATOR")
+    public UniversityDtoDetailed addNewUniversity(UniversityDtoFormCreate form) {
+        if (universityRepository.existsByNameOrShortName(form.getName(), form.getShortName())) {
+            throw new UniversityException(UniversityExceptionType.NAME_TAKEN);
+        }
+
+        User creator = userRepository.findByUsername(form.getCreatorUsername())
+                .orElseThrow(() -> {
+                    throw new BadRequestException("Not found user");
+                });
+
+        if (!securityService.hasPermissionsToUser(creator)) {
+            throw new ForbiddenException();
+        }
+
+        University newUniversity = form.toUniversity(creator);
+        if (!securityService.hasPermissionsToUniversity(newUniversity)) {
+            throw new ForbiddenException();
+        }
+
+        return new UniversityDtoDetailed(universityRepository.save(newUniversity));
+    }
+
+    @Secured("ROLE_MODERATOR")
+    public void update(Long id, UniversityDtoFormUpdate form) {
+        University university = universityRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!securityService.hasPermissionsToUniversity(university)) {
+            throw new ForbiddenException();
+        }
+
+        form.updateUniversity(university);
+        universityRepository.save(university);
+    }
+
     @Transactional
+    @Secured("ROLE_ADMIN") // TODO: Remove method?
     public UniversityDtoDetailed enrollUsersToUniversity(Long universityId, Long userId) {
 
         University university = universityRepository.findById(universityId).orElseThrow(NotFoundException::new);
@@ -47,6 +95,7 @@ public class UniversityService {
         return new UniversityDtoDetailed(result);
     }
 
+    @Secured("ROLE_ADMIN") // TODO: Remove method?
     public University connectMainPageToUniversity(Long universityId, Long pageId) {
 
         University university = universityRepository.findById(universityId).orElseThrow(NotFoundException::new);
@@ -56,58 +105,27 @@ public class UniversityService {
         return universityRepository.save(university);
     }
 
-
-    public UniversityDtoDetailed getUniversity(Long id) {
-        return universityRepository.findById(id).map(UniversityDtoDetailed::new).orElseThrow(NotFoundException::new);
-
-    }
-
-    public UniversityDtoDetailed addNewUniversity(UniversityDtoForm form) {
-        return new UniversityDtoDetailed(universityRepository.save(formToUniversity(form)));
-    }
-
-    public University formToUniversity(UniversityDtoForm form) {
-        University university = new University();
-
-        if (universityRepository.existsByNameOrShortName(form.getName(), form.getShortName())) {
-            throw new UniversityException(UniversityExceptionType.NAME_TAKEN);
-        }
-
-        university.setName(form.getName());
-        university.setShortName(form.getShortName());
-        university.setHidden(true);
-
-        Page page = new Page();
-        page.setTitle(university.getName());
-        page.setDescription("Short description about university.");
-        page.setContent("Content for university main page.");
-        page.setHidden(true);
-        page.setUniversity(university);
-
-        User creator = userRepository.findByUsername(form.getCreatorUsername())
-                .orElseThrow(() -> {
-                    throw new BadRequestException("Not found user");
-                });
-
-        page.setCreator(creator);
-
-        university.setMainPage(page);
-
-        return university;
-    }
-
+    @Secured("ROLE_MODERATOR")
     public void modifyHiddenField(Long id, boolean hidden) {
         University university = universityRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!securityService.hasPermissionsToUniversity(university)) {
+            throw new ForbiddenException();
+        }
+
         university.setHidden(hidden);
         universityRepository.save(university);
     }
 
+    @Secured("ROLE_MODERATOR")
     public void deleteUniversity(Long id) {
         University university = universityRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!securityService.hasPermissionsToUniversity(university)) {
+            throw new ForbiddenException();
+        }
+
         validateForDelete(university);
         universityRepository.delete(university);
     }
-
     private void validateForDelete(University university) {
         if (!university.isHidden()) {
             throw new UniversityException(UniversityExceptionType.UNIVERSITY_IS_NOT_HIDDEN);
@@ -121,14 +139,5 @@ public class UniversityService {
                 throw new UniversityException(UniversityExceptionType.ACTIVE_USER_EXISTS);
             }
         }
-    }
-
-    public UniversityDtoDetailed save(UniversityDtoForm form){
-        return new UniversityDtoDetailed(universityRepository.save(formToUniversity(form)));
-    }
-    public void update(Long id, UniversityDtoForm form) {
-        University university = universityRepository.findById(id).orElseThrow(NotFoundException::new);
-        university.updateFromEditedForm(form);
-        universityRepository.save(university);
     }
 }
